[C] Move ProgressBar to /loading
