+++ components/atomic/loading/ProgressBar/ProgressBar.styles.ts
